Lots of changes, MESSY stage

+++ eal.sh
@@ -1,11 +1,12 @@
 #!/bin/bash
 cd $( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )
 source ceal.sh
-clear
-
-mess -t "Ewancoder Arch Linux installation script\nVersion $version"
-mess -w "Before proceeding, MAKE SURE that\n\t1) You have changed all constants in 'ceal.sh' file\n\t2) You have FORMATTED your partitions as needed (fdisk + mkfs.ext4) and put them into 'ceal.sh' file"
-source ceal.sh
+if ! [ "$1" == "--hide" ]; then
+    clear
+    mess -t "Ewancoder Arch Linux installation script\nVersion $version"
+    mess -w "Before proceeding, MAKE SURE that\n\t1) You have changed all constants in 'ceal.sh' file\n\t2) You have FORMATTED your partitions as needed (fdisk + mkfs.ext4) and put them into 'ceal.sh' file"
+    source ceal.sh
+fi
 
 mess -t "Mount all partitions and create fstab"
 mess "Create local fstab file (prepare)"
@@ -19,7 +20,7 @@ for (( i = 0; i < ${#device[@]}; i++ )); do
     echo -e "\n# ${description[$i]} partition\n${device[$i]}\t${mount[$i]}\t${type[$i]}\t${option[$i]}\t${dump[$i]}\t${pass[$i]}" >> fstab
 done
 
-mess -t "Form mirrorlist"
+mess -t "Form mirrorlist & update pacman"
 for i in "${mirror[@]}"
 do
     mess "Place $i in mirrorlist"
@@ -34,15 +35,13 @@ pacman-key --init
 pacman-key --populate archlinux
 pacman-key --keyserver hkp://pgp.mit.edu -r B02854ED753E0F1F
 
-mess -t "Install & setup system"
+mess -t "Install system"
 mess "Install base-system"
 pacstrap /mnt base base-devel
 mess "Move previously created fstab to /mnt/etc/fstab"
 mv fstab /mnt/etc/fstab
-mess "Set hostname ($hostname)"
-echo $hostname > /mnt/etc/hostname
 
-mess -t "CHROOT to system"
+mess -t "Chroot to system"
 mess "Create root folder (just in case)"
 mkdir -p /mnt/root
 mess "Prepare eal-chroot.sh chroot script"
@@ -50,6 +49,9 @@ echo '
 #!/bin/bash
 source /root/ceal.sh
 
+mess -t "Setup hostname & timezone"
+mess "Set hostname ($hostname)"
+echo $hostname > /etc/hostname
 mess "Set local timezone ($timezone)"
 ln -fs /usr/share/zoneinfo/$timezone /etc/localtime
 
@@ -63,6 +65,7 @@ locale-gen
 mess "Set font as $font"
 setfont $font
 
+mess -t "Install grub"
 mess "Install grub to /boot"
 pacman -S --noconfirm grub
 mess "Install grub bootloader to $mbr mbr"
@@ -71,29 +74,38 @@ mess "Install os-prober"
 pacman -S --noconfirm os-prober
 mess "Make grub config"
 grub-mkconfig -o /boot/grub/grub.cfg
-
-mess -p "Setup ROOT password"
-passwd
 ' > /mnt/root/eal-chroot.sh
 mess "Add peal.sh to eal-chroot script"
 cat peal.sh >> /mnt/root/eal-chroot.sh
-mess "Add 'exit chroot' message to the end of eal-chroot script"
+mess "Add passwords setup & exit chroot sections"
 echo '
-mess "Remove all scripts"
-rm /root/{eal-chroot,ceal,peal}.sh
+mess -t "Setup all passwords"
+mess -p "Setup ROOT password"
+passwd
+for i in ${user[@]}; do
+    mess -p "Setup user ($i) password"
+    passwd $i
+done
 
-mess "Exit chroot"
+mess -t "Finish installation"
+mess "Remove all scripts"
+rm /root/{eal-chroot,ceal}.sh
+mess "Exit chroot (installed system -> live-cd)"
 exit
 ' >> /mnt/root/eal-chroot.sh
 mess "Set executable flag for chroot script"
 chmod +x /mnt/root/eal-chroot.sh
-mess "Copy {ceal,peal}.sh to /mnt/root"
-cp {ceal,peal}.sh /mnt/root/
+mess "Copy ceal.sh to /mnt/root"
+cp ceal.sh /mnt/root/
 mess "Go to arch-chroot"
 arch-chroot /mnt /root/eal-chroot.sh
 
-mess -t "Finish installation"
-mess "Unmount all within /mnt"
+mess "Unmount all within /mnt (unmount installed system)"
 umount -R /mnt
-mess "Exiting chroot"
-exit
+if [ "$1" == "--hide" ]; then
+    mess "Exiting chroot (live-cd -> host system)"
+    exit
+else
+    mess -w "This is it. Your system is now installed [REBOOT]"
+    reboot
+fi
